tree: a new design of `PositionTree`
