+++ Cargo.toml
@@ -10,15 +10,23 @@ rust-version = "1.65.0"
 
 [features]
 default = ["std", "serde"]
-alloc = ["smol_str"]
+alloc = [
+    "smol_str",
+    "hashbrown",
+    "serde?/alloc",
+]
 std = [
     "alloc",
     "thiserror",
     "serde?/std",
     "smol_str?/std",
-    "im",
 ]
-serde = ["dep:serde", "serde_with", "smol_str?/serde"]
+serde = [
+    "dep:serde",
+    "serde_with",
+    "smol_str?/serde",
+    "hashbrown?/serde",
+]
 
 [dependencies]
 smol_str = { version = "0.1.23", default-features = false, optional = true }
@@ -26,7 +34,7 @@ num-traits = { version = "0.2.14", default_features = false }
 serde = { version = "1", features = ["derive"], optional = true, default_features = false }
 serde_with = { version = "2.1", features = ["macros"], optional = true, default_features = false }
 thiserror = { version = "1.0", optional = true }
-im = { version = "15.1.0", optional = true }
+hashbrown = { version = "0.13.1", optional = true }
 
 [dev-dependencies]
 rust_decimal = "1.26.1"
